<Search/> moved to 'Navigation' folder
